RFCT: move zoom scale and markers to the relevant widgets for readability
