+++ frontend/src/components/widgets/image-holder/constants/zoom-scale.js
@@ -1,0 +1,1 @@
+export const ZOOM_SCALE = { scroll: 10, focus: 7 };
