+++ frontend/src/components/widgets/mark-canvas/constants/markers.js
@@ -1,0 +1,5 @@
+export const MARKERS = {
+  temp: { id: 0, name: "temp", color: "transparent", radius: 3 },
+  static: { id: 0, name: "static", color: "#ffff00", radius: 2 },
+  zoom: { id: 0, name: "zoom", color: "chartreuse", radius: 10 },
+};
