+++ frontend/src/components/widgets/mark-canvas/hooks/useDotMarker.js
@@ -1,6 +1,6 @@
 import { useMemo } from "react";
 
-import { MARKERS } from "@/core/constants";
+import { MARKERS } from "../constants/markers";
 
 const useDotMarker = (imageSize, marks, coordinates) => {
   const generateCircles = useMemo(() => {
@@ -9,18 +9,23 @@ const useDotMarker = (imageSize, marks, coordinates) => {
     const marksMap = new Map();
     marks.forEach((mark) => {
       marksMap.set(mark.fileName, mark);
+      marksMap.set(mark.file_name + MARKERS.zoom.name, mark);
     });
 
     return (coordinates || []).map((file, index) => {
       const dx = Math.round(file.norm_x_center * imageSize.width * 100) / 100;
       const dy = Math.round(file.norm_y_center * imageSize.height * 100) / 100;
 
+      const stored_mark =
+        marksMap.get(file.file_name) ||
+        marksMap.get(file.file_name + MARKERS.zoom.name);
+
       return {
         id: index,
         cx: dx,
         cy: dy,
-        r: MARKERS.temp.radius,
-        color: MARKERS.static.color,
+        r: stored_mark?.marker.radius || MARKERS.static.radius,
+        color: stored_mark?.marker.color || MARKERS.static.color,
       };
     });
   }, [coordinates, marks, imageSize]);
