--- frontend/src/core/constants.js
@@ -1,8 +1,0 @@
-const MARKERS = {
-  temp: { id: 0, name: "temp", color: "transparent", radius: 3 },
-  static: { id: 0, name: "static", color: "#ffff00", radius: 1 },
-  zoom: { id: 0, name: "zoom", color: "chartreuse", radius: 10 },
-};
-const ZOOM_SCALE = { scroll: 10, focus: 7 };
-
-export { MARKERS, ZOOM_SCALE };
